+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 - Docs: Update new-address-api.md (#360)
 - feat: worker 增加 `ADMIN_USER_ROLE` 配置, 用于配置管理员用户角色，此角色的用户可访问 admin 管理页面 (#363)
-- feat: worker 增加 `SHOW_GITHUB` 配置, 用于配置是否显示 github 链接
+- feat: worker 增加 `DISABLE_SHOW_GITHUB` 配置, 用于配置是否显示 github 链接
+- feat: worker 增加 `NO_LIMIT_SEND_ROLE` 配置, 用于配置可以无限发送邮件的角色
 
 ## v0.6.1